m42634/syntax: represent quant step size in 32 bits rather than 16
This commit fixes a syntax issue where attribute quantisation step sizes
greater than 2^{16}-1 could not be correctly represented.

Extracted-from: http://wg11.sc29.org/svn/repos/MPEG-I/Part5-PointCloudCompression/CE/CE0/CE0.1/trunk@1357
Reintegrated-by: dflynn@blackberry.com

+++ tmc3/PCCTMC3Decoder.h
@@ -392,8 +392,8 @@ class PCCTMC3Decoder3 {
 
     quantizationSteps.resize(levelOfDetailCount);
     for (size_t lodIndex = 0; lodIndex < levelOfDetailCount; ++lodIndex) {
-      uint16_t qs = 0;
-      PCCReadFromBuffer<uint16_t>(bitstream.buffer, qs, bitstream.size);
+      uint32_t qs = 0;
+      PCCReadFromBuffer<uint32_t>(bitstream.buffer, qs, bitstream.size);
       quantizationSteps[lodIndex] = qs;
     }
